requirements.txt
flask
pymupdf
sentence-transformers
faiss-cpu
torch 
mistralai
python-dotenv